add user model, repo, endpoints, migrations, rest client file; add ApiResponse Responder type to handle errors and serializable trait implementing data implicitly and consistently; add opionated rustfmt.toml; add identity and session middleware using custom postgres session store to be used as authentication of users later on;

+++ api/lib/src/todo_repository/mod.rs
@@ -9,9 +9,9 @@ pub use postgres_todo_repository::PostgresTodoRepository;
 
 #[async_trait::async_trait]
 pub trait TodoRepository: Send + Sync + 'static {
-    async fn get_todos(&self) -> TodoResult<Vec<Todo>>;
-    async fn get_todo(&self, id: &i32) -> TodoResult<Todo>;
-    async fn create_todo(&self, create_todo: &CreateTodo) -> TodoResult<Todo>;
-    async fn update_todo(&self, update_todo: &UpdateTodo) -> TodoResult<Todo>;
-    async fn delete_todo(&self, id: &i32) -> TodoResult<i32>;
+  async fn get_todos(&self) -> TodoResult<Vec<Todo>>;
+  async fn get_todo(&self, id: &i64) -> TodoResult<Todo>;
+  async fn create_todo(&self, create_todo: &CreateTodo) -> TodoResult<Todo>;
+  async fn update_todo(&self, update_todo: &UpdateTodo) -> TodoResult<Todo>;
+  async fn delete_todo(&self, id: &i64) -> TodoResult<i64>;
 }
